backend/requirements.txt
fastapi
uvicorn
PyMuPDF
python-multipart
pydantic
python-dotenv
